requirements.txt
Pillow==11.3.0
tox==4.30.2
pytest==8.4.2